% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_data}
\alias{as_data}
\title{Tells obsplot that a single value is a data channel, and must be converted to
an Array in JavaScript}
\usage{
as_data(x)
}
\arguments{
\item{x}{the object to be considered as a data channel}
}
\description{
Tells obsplot that a single value is a data channel, and must be converted to
an Array in JavaScript
}
