% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{barley}
\alias{barley}
\title{Sample data from Observable Plot documentation examples}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 120 rows and 4 columns.
}
\source{
\url{https://observablehq.com/@observablehq/plot-facets}
}
\usage{
barley
}
\description{
Sample data from Observable Plot documentation examples
}
\keyword{datasets}
