% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet.R
\name{facet}
\alias{facet}
\title{Add a facet to an obsplot}
\usage{
facet(g, data = NULL, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{data}{optional data}

\item{...}{options passed to define the facet}
}
\description{
Add a facet to an obsplot
}
