% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_area}
\alias{mark_area}
\alias{mark_areaX}
\alias{mark_areaY}
\alias{mark_barX}
\alias{mark_barY}
\alias{mark_cell}
\alias{mark_cellX}
\alias{mark_cellY}
\alias{mark_dot}
\alias{mark_dotX}
\alias{mark_dotY}
\alias{mark_line}
\alias{mark_lineX}
\alias{mark_lineY}
\alias{mark_link}
\alias{mark_rect}
\alias{mark_rectX}
\alias{mark_rectY}
\alias{mark_ruleX}
\alias{mark_ruleY}
\alias{mark_text}
\alias{mark_textX}
\alias{mark_textY}
\alias{mark_tickX}
\alias{mark_tickY}
\alias{mark_image}
\title{Add a mark to an obsplot}
\usage{
mark_area(g, ...)

mark_areaX(g, ...)

mark_areaY(g, ...)

mark_barX(g, ...)

mark_barY(g, ...)

mark_cell(g, ...)

mark_cellX(g, ...)

mark_cellY(g, ...)

mark_dot(g, ...)

mark_dotX(g, ...)

mark_dotY(g, ...)

mark_line(g, ...)

mark_lineX(g, ...)

mark_lineY(g, ...)

mark_link(g, ...)

mark_rect(g, ...)

mark_rectX(g, ...)

mark_rectY(g, ...)

mark_ruleX(g, ...)

mark_ruleY(g, ...)

mark_text(g, ...)

mark_textX(g, ...)

mark_textY(g, ...)

mark_tickX(g, ...)

mark_tickY(g, ...)

mark_image(g, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{...}{named options passed to define the mark

... arguments can accept at most two unnamed arguments : a data object and a call
to a transform function.}
}
\description{
Add a mark to an obsplot
}
