% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_frame}
\alias{mark_frame}
\title{Add a decoration mark to an obsplot}
\usage{
mark_frame(g, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{...}{options passed to define the mark}
}
\description{
Add a decoration mark to an obsplot
}
