% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_function}
\alias{mark_function}
\title{Add a render function mark to an obsplot}
\usage{
mark_function(g, f, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{f}{render function, defined with \code{htmlwidgets::JS()}}

\item{...}{options passed to define the mark}
}
\description{
Add a render function mark to an obsplot
}
