% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marks.R
\name{mark_svg}
\alias{mark_svg}
\title{Add an SVG element to an obsplot}
\usage{
mark_svg(g, svg, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{svg}{SVG to add, as a character string}

\item{...}{options passed to define the mark}
}
\description{
Add an SVG element to an obsplot
}
