% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mobydick1}
\alias{mobydick1}
\title{Sample data from Observable Plot documentation examples}
\format{
An object of class \code{character} of length 9622.
}
\source{
\url{https://observablehq.com/@observablehq/plot-group}
}
\usage{
mobydick1
}
\description{
Sample data from Observable Plot documentation examples
}
\keyword{datasets}
