% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsplot.R
\name{obsplot}
\alias{obsplot}
\title{Create a new obsplot plot}
\usage{
obsplot(data = NULL, width = NULL, height = NULL, elementId = NULL, ...)
}
\arguments{
\item{data}{optional data.frame as primary data source}

\item{width}{plot width}

\item{height}{plot height}

\item{elementId}{optional custom element id, passed to createWidget()}

\item{...}{named options passed to Observable Plot}
}
\description{
Create a new obsplot plot
}
