% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{povcalnet}
\alias{povcalnet}
\title{Sample data from Observable Plot documentation examples}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 173 rows and 32 columns.
}
\source{
\url{https://observablehq.com/@observablehq/plot-rect}
}
\usage{
povcalnet
}
\description{
Sample data from Observable Plot documentation examples
}
\keyword{datasets}
