% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sfcases}
\alias{sfcases}
\title{Sample data from Observable Plot changelog}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1797 rows and 5 columns.
}
\source{
\url{https://observablehq.com/@visnup/sf-covid-data}
}
\usage{
sfcases
}
\description{
Sample data from Observable Plot changelog
}
\keyword{datasets}
