% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opts.R
\name{style}
\alias{style}
\title{Add or modify style options to an obsplot}
\usage{
style(g, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{...}{named style options}
}
\description{
Add or modify style options to an obsplot
}
