% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_normalizeX}
\alias{transform_normalizeX}
\alias{transform_normalizeY}
\alias{transform_windowX}
\alias{transform_windowY}
\alias{transform_selectFirst}
\alias{transform_selectLast}
\alias{transform_selectMinX}
\alias{transform_selectMaxX}
\alias{transform_selectMinY}
\alias{transform_selectMaxY}
\alias{transform_stackY}
\alias{transform_stackY1}
\alias{transform_stackY2}
\alias{transform_stackX}
\alias{transform_stackX1}
\alias{transform_stackX2}
\title{Define an Observable Plot transform with options only}
\usage{
transform_normalizeX(options)

transform_normalizeY(options)

transform_windowX(options)

transform_windowY(options)

transform_selectFirst(options)

transform_selectLast(options)

transform_selectMinX(options)

transform_selectMaxX(options)

transform_selectMinY(options)

transform_selectMaxY(options)

transform_stackY(options)

transform_stackY1(options)

transform_stackY2(options)

transform_stackX(options)

transform_stackX1(options)

transform_stackX2(options)
}
\arguments{
\item{options}{a named list of options}
}
\description{
Define an Observable Plot transform with options only
}
