% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_selectFirst}
\alias{transform_selectFirst}
\alias{transform_selectLast}
\alias{transform_selectMinX}
\alias{transform_selectMaxX}
\alias{transform_selectMinY}
\alias{transform_selectMaxY}
\title{Define an Observable Plot transform options only}
\usage{
transform_selectFirst(...)

transform_selectLast(...)

transform_selectMinX(...)

transform_selectMaxX(...)

transform_selectMinY(...)

transform_selectMaxY(...)
}
\arguments{
\item{...}{transform options}
}
\description{
Define an Observable Plot transform options only
}
