% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_windowX}
\alias{transform_windowX}
\alias{transform_windowY}
\alias{transform_stackY}
\alias{transform_stackY1}
\alias{transform_stackY2}
\alias{transform_stackX}
\alias{transform_stackX1}
\alias{transform_stackX2}
\title{Define an Observable Plot transform with two options objects}
\usage{
transform_windowX(spec = NULL, ...)

transform_windowY(spec = NULL, ...)

transform_stackY(spec = NULL, ...)

transform_stackY1(spec = NULL, ...)

transform_stackY2(spec = NULL, ...)

transform_stackX(spec = NULL, ...)

transform_stackX1(spec = NULL, ...)

transform_stackX2(spec = NULL, ...)
}
\arguments{
\item{spec}{specific options}

\item{...}{transform options}
}
\description{
Define an Observable Plot transform with two options objects
}
