% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{trends2020}
\alias{trends2020}
\title{Sample data from Observable Plot changelog}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 3392 rows and 7 columns.
}
\source{
\url{https://observablehq.com/@ben-tanen/the-relentless-2020-news-cycle-in-one-chart}
}
\usage{
trends2020
}
\description{
Sample data from Observable Plot changelog
}
\keyword{datasets}
