% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsplot.R
\name{obsplot}
\alias{obsplot}
\title{Create a new obsplot}
\usage{
obsplot(
  data = NULL,
  width = "auto",
  height = "auto",
  menu = getOption("obsplot_menu"),
  elementId = NULL,
  ...
)
}
\arguments{
\item{data}{optional data.frame as primary data source}

\item{width}{plot width. Use NULL to let Plot decide, or "auto" to let htmlwidgets decide}

\item{height}{plot height. Use NULL to let Plot decide, or "auto" to let htmlwidgets decide}

\item{menu}{if TRUE, display a gear menu aside the plot with additional features. Can be set globally with \code{options("obsplot_menu")}}

\item{elementId}{optional custom element id, passed to createWidget()}

\item{...}{named options passed to Observable Plot}
}
\description{
Create a new obsplot
}
