% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color}
\alias{scale_color}
\alias{scale_opacity}
\alias{scale_radius}
\alias{scale_x}
\alias{scale_y}
\title{Add or modify scale options to an obsplot}
\usage{
scale_color(g, ...)

scale_opacity(g, ...)

scale_radius(g, ...)

scale_x(g, ...)

scale_y(g, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{...}{named scale options}
}
\description{
Add or modify scale options to an obsplot
}
