% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color}
\alias{scale_color}
\alias{scale_opacity}
\alias{scale_r}
\alias{scale_x}
\alias{scale_y}
\alias{scale_fx}
\alias{scale_fy}
\title{Define an obsplot scale options}
\usage{
scale_color(g, ...)

scale_opacity(g, ...)

scale_r(g, ...)

scale_x(g, ...)

scale_y(g, ...)

scale_fx(g, ...)

scale_fy(g, ...)
}
\arguments{
\item{g}{an obsplot object}

\item{...}{named scale options}
}
\description{
Define an obsplot scale options
}
