% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_bin}
\alias{transform_bin}
\alias{transform_binX}
\alias{transform_binY}
\alias{transform_group}
\alias{transform_groupX}
\alias{transform_groupY}
\alias{transform_groupZ}
\alias{transform_map}
\title{Define an Observable Plot transform with outputs and options}
\usage{
transform_bin(outputs, options)

transform_binX(outputs, options)

transform_binY(outputs, options)

transform_group(outputs, options)

transform_groupX(outputs, options)

transform_groupY(outputs, options)

transform_groupZ(outputs, options)

transform_map(outputs, options)
}
\arguments{
\item{outputs}{a named list of outputs}

\item{options}{a named list of options}
}
\description{
Define an Observable Plot transform with outputs and options
}
