% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_bin}
\alias{transform_bin}
\alias{transform_binX}
\alias{transform_binY}
\alias{transform_group}
\alias{transform_groupX}
\alias{transform_groupY}
\alias{transform_groupZ}
\alias{transform_map}
\title{Define an Observable Plot transform with outputs and options}
\usage{
transform_bin(...)

transform_binX(...)

transform_binY(...)

transform_group(...)

transform_groupX(...)

transform_groupY(...)

transform_groupZ(...)

transform_map(...)
}
\arguments{
\item{...}{transform outputs and options}
}
\description{
Define an Observable Plot transform with outputs and options
}
\details{
transform output is given as first named argument of the function. If there are several outputs, pass a named list as first argument.
}
