% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_mapX}
\alias{transform_mapX}
\alias{transform_mapY}
\title{Define an Observable Plot transform with map and options}
\usage{
transform_mapX(map, options)

transform_mapY(map, options)
}
\arguments{
\item{map}{a map method name}

\item{options}{a named list of options}
}
\description{
Define an Observable Plot transform with map and options
}
