% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transforms.R
\name{transform_normalizeX}
\alias{transform_normalizeX}
\alias{transform_normalizeY}
\title{Define an Observable Plot transform with basis and options}
\usage{
transform_normalizeX(basis = NULL, ...)

transform_normalizeY(basis = NULL, ...)
}
\arguments{
\item{basis}{a basis method name}

\item{...}{transform options}
}
\description{
Define an Observable Plot transform with basis and options
}
